decouple DTOs from domain and improve data transfer design

+++ aliceGlow/src/main/java/aliceGlow/example/aliceGlow/domain/Product.java
@@ -6,6 +6,8 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "products")
 public class Product {
@@ -19,11 +21,11 @@ public class Product {
 
     @NotNull
     @DecimalMin("0.01")
-    private Double price;
+    private BigDecimal price;
 
     @NotNull
     @Min(0)
-    private Long stock;
+    private Integer stock;
 
     public Long getId(){return id;}
     public void setId(Long id){this.id = id;}
@@ -31,9 +33,9 @@ public class Product {
     public String getName(){return name;}
     public void setName(String name){this.name = name;}
 
-    public Double getPrice(){return price;}
-    public void setPrice(Double price){this.price = price;}
+    public BigDecimal getPrice(){return price;}
+    public void setPrice(BigDecimal price){this.price = price;}
 
-    public Long getStock(){return stock;}
-    public void setStock(Long stock){this.stock = stock;}
+    public Integer getStock(){return stock;}
+    public void setStock(Integer stock){this.stock = stock;}
 }
